pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webpconverter"
version = "0.1.0"
description = "CLI tool to convert PNG, JPEG and SVG assets to optimized WebP images"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "click>=8.1",
    "Pillow>=10.0",
    "cairosvg>=2.7",
    "rich>=13.0",
    "PyYAML>=6.0",
]

[project.scripts]
webpconverter = "webpconverter.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["webpconverter*"]
